Renamed momentmanager -> datamanagers
